move verify proxy to folder
